Refactored tests to allow running on arbitrary operating systems with the RUN_TESTS flag set to true

+++ test/BeagleBone/GPIO_test.jl
@@ -7,40 +7,40 @@ using Base.Test
 
     @testset "Inialization/Termination" begin
         # Initialize three devices
-        initdev("gpio", 1)
+        initdev("gpio", Int32(1))
         @test sum(listdev()) == 1
 
-        initdev("gpio", 3)
+        initdev("gpio", Int32(3))
         @test sum(listdev()) == 2
 
-        initdev("gpio", 5)
+        initdev("gpio", Int32(5))
         @test sum(listdev()) == 3
 
         #printdev("gpio", 3)
 
         # Attempt to initialize a device which has already been initialized
-        @test_throws ErrorException initdev("gpio", 1)
-        @test_throws ErrorException initdev("gpio", 3)
-        @test_throws ErrorException initdev("gpio", 5)
+        @test_throws ErrorException initdev("gpio", Int32(1))
+        @test_throws ErrorException initdev("gpio", Int32(3))
+        @test_throws ErrorException initdev("gpio", Int32(5))
 
         # Attempt to initialize a device with a very high index (no matching channel)
-        @test_throws ErrorException initdev("gpio", 1000)
+        @test_throws ErrorException initdev("gpio", Int32(1000))
 
         # Attempt to remove devices which have not been initialized
-        @test_throws ErrorException closedev("gpio", 2)
-        @test_throws ErrorException closedev("gpio", 4)
-        @test_throws ErrorException closedev("gpio", 6)
+        @test_throws ErrorException closedev("gpio", Int32(2))
+        @test_throws ErrorException closedev("gpio", Int32(4))
+        @test_throws ErrorException closedev("gpio", Int32(6))
 
         #printdev("gpio", 3)
 
         # Remove devices from TOC
-        closedev("gpio", 1)
+        closedev("gpio", Int32(1))
         @test sum(listdev()) == 2
 
-        closedev("gpio", 3)
+        closedev("gpio", Int32(3))
         @test sum(listdev()) == 1
 
-        closedev("gpio", 5)
+        closedev("gpio", Int32(5))
         @test sum(listdev()) == 0
 
     end
@@ -48,45 +48,45 @@ using Base.Test
     @testset "Read/Write" begin
 
         # Fixture
-        device = initdev("gpio", 1)
+        device = initdev("gpio", Int32(1))
 
         # Test that an exception is thrown when an invalid operation is given
         # supported operations are 1,2,3
-        @test_throws ErrorException write!(device, (0, "something"))
-        @test_throws ErrorException write!(device, (4, "something"))
+        @test_throws ErrorException write!(device, (Int32(0), "something"))
+        @test_throws ErrorException write!(device, (Int32(4), "something"))
 
         # Test that exceptions are thrown for each individual operation
-        @test_throws ErrorException write!(device, (1, "bad_entry"))
-        @test_throws ErrorException write!(device, (2, "bad_entry"))
-        @test_throws ErrorException write!(device, (3, "bad_entry"))
+        @test_throws ErrorException write!(device, (Int32(1), "bad_entry"))
+        @test_throws ErrorException write!(device, (Int32(2), "bad_entry"))
+        @test_throws ErrorException write!(device, (Int32(3), "bad_entry"))
 
         # Test operation 1
-        write!(device, (1, "1"))
-        @test read(device, 1) == "1"
-        write!(device, (1, "0"))
-        @test read(device, 1) == "0"
-        write!(device, (1, "1"))
-        @test read(device, 1) == "1"
-        write!(device, (1, "0"))
-        @test read(device, 1) == "0"
+        write!(device, (Int32(1), "1"))
+        @test read(device, Int32(1)) == "1"
+        write!(device, (Int32(1), "0"))
+        @test read(device, Int32(1)) == "0"
+        write!(device, (Int32(1), "1"))
+        @test read(device, Int32(1)) == "1"
+        write!(device, (Int32(1), "0"))
+        @test read(device, Int32(1)) == "0"
 
-        write!(device, (2, "in"))
-        @test read(device, 2) == "in"
-        write!(device, (2, "out"))
-        @test read(device, 2) == "out"
-        write!(device, (2, "in"))
-        @test read(device, 2) == "in"
-        write!(device, (2, "out"))
-        @test read(device, 2) == "out"
+        write!(device, (Int32(2), "in"))
+        @test read(device, Int32(2)) == "in"
+        write!(device, (Int32(2), "out"))
+        @test read(device, Int32(2)) == "out"
+        write!(device, (Int32(2), "in"))
+        @test read(device, Int32(2)) == "in"
+        write!(device, (Int32(2), "out"))
+        @test read(device, Int32(2)) == "out"
 
         # Test operation 3
-        @test_throws ErrorException write!(device, (3, "none"))
-        @test_throws ErrorException write!(device, (3, "rising"))
-        @test_throws ErrorException write!(device, (3, "falling"))
-        @test_throws ErrorException write!(device, (3, "both"))
+        @test_throws ErrorException write!(device, (Int32(3), "none"))
+        @test_throws ErrorException write!(device, (Int32(3), "rising"))
+        @test_throws ErrorException write!(device, (Int32(3), "falling"))
+        @test_throws ErrorException write!(device, (Int32(3), "both"))
 
         # Close Gpio
-        closedev("gpio", 1)
+        closedev("gpio", Int32(1))
      end
 
     @testset "All channels" begin
@@ -96,9 +96,9 @@ using Base.Test
         # Configure the GPIO for output usage
         devices = []
         for ii = 1:length(gpio_channels)
-            device = initdev("gpio", ii)
+            device = initdev("gpio", Int32(ii))
             # Operation 2 -> in/out, set out
-            write!(device, (2, "out"))
+            write!(device, (Int32(2), "out"))
             # Append to list
             append!(devices, [device])
         end
@@ -107,15 +107,15 @@ using Base.Test
         for i = 1:10
             for ii = 1:length(gpio_channels)
                 state = "$(i%2)"
-                write!(devices[ii], (1, state))
-                @test read(devices[ii], 1) == state
+                write!(devices[ii], (Int32(1), state))
+                @test read(devices[ii], Int32(1)) == state
             end
             sleep(0.10)
         end
 
         # Closes all devices
         for ii = 1:length(gpio_channels)
-            closedev("gpio", ii)
+            closedev("gpio", Int32(ii))
         end
     end
 end
